4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -40,6 +40,15 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    
+    if ( blog.user.toString() !== user._id.toString()) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    } 
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
